Add backend changes

+++ src/sentry/api/endpoints/user_authenticator_details.py
@@ -134,6 +134,14 @@ class UserAuthenticatorDetailsEndpoint(UserEndpoint):
             return Response(status=status.HTTP_204_NO_CONTENT)
 
         with transaction.atomic():
+            # if the user's organization requires 2fa,
+            # don't delete the last auth method
+            enrolled_methods = Authenticator.objects.all_interfaces_for_user(user, ignore_backup=True)
+            last_authenticator = len(enrolled_methods) == 1
+            require_2fa = user.get_orgs_require_2fa()
+
+            if require_2fa and last_authenticator:
+                return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
             authenticator.delete()
 
             # if we delete an actual authenticator and all that
